requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
streamlit==1.29.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
openai==1.6.1
tiktoken==0.5.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
